feat: add incident notification functionality for workspace creation in Slack
